+++ src/Api/Leagues.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Humble23\CartolaFcClient\Api;
+
+use Humble23\CartolaFcClient\Api\Api;
+
+class Leagues extends Api
+{
+    public function all($leagueName = '')
+    {
+        return $this->get('/ligas', [
+            'q' => $leagueName,
+        ]);
+    }
+}