+++ CHANGELOG.md
@@ -8,7 +8,9 @@ All notable changes to `cartola-fc-api` will be documented in this file.
 - Added Guzzle Http Client
 - Added Classe with market endpoints
 
-## 0.2.0 - 2022-12-31
+## 0.4.0 - 2022-12-31
 
-- Added Classe with team endpoints
+- Added Classe with teams endpoints
 - Added Classe to transform api response
+- Added Classe with game endpoints
+- Added Classe with leagues endpoints